[DO NOT MERGE!!!] LPS-122191 Fix CI
Note that this fixes the problem at https://github.com/liferay-appsec/liferay-portal/pull/143#issuecomment-712284026 but at the cost of a major change in portal-kernel.

The diagnostic of the problem is explained in the comments of the same PR of the error, starting at https://github.com/liferay-appsec/liferay-portal/pull/143#issuecomment-713662990.