LPD-31793 It is unexpected that assigning a policy will leave a mark on the user itself.
It should be reversible. Also, multiple policies can be associated directly or indirectly. After LPD-29411, this is cleanly dealt with in the auth process where the applicable policy is already resolved.

Reverts "LPS-76504 portal-impl - UserLocalServiceImpl.java - when settings a p******d policy to a user, check to see if the p******d policy allows changing the p******d.  If not, make sure that the user is not forced to reset their p******d by setting the "p******dReset" field to false"

This reverts commit cb6629b3