LPD-34973 The 'value' attribute does not get incremented with auto fields, but the name can.  The name attribute is currently set to a static, non-incrementable string which is used to determine which attribute was seleted for matching.  Let's move this to a hidden field, and instead allow the name to be incrementable for the radio field.  This pattern closely follows how we determine the userIdentifierExpressionPrefix attribute.
Because the front-end can no longer retrieve all radio fields using the previous name string, let's change the query selectors to look for a radio type instead.

Lastly, since the name and id attributes are identical, and the name value is used for the id if id isn't present, we can safely remove the id attribute.