LPD-67475 SF Consistency
# breaking

## What modules/apps/oauth2-provider/oauth2-provider-api/src/main/java/com/liferay/oauth2/provider/constants/OAuth2ProviderActionKeys.java

Renamed action keys names, removed the namespace

## Why

The namespace is not needed for the constants in this class to follow same pattern as in PushNotificationsActionKeys.java.

----

# breaking

## What modules/apps/oauth2-provider/oauth2-provider-api/src/main/java/com/liferay/oauth2/provider/constants/OAuth2ProviderConstants.java

Renamed keys names, added namespace

## Why

The namespace is needed for the constants in this class to follow same pattern as in AccessibilitySettingConstants or AccountConstants.java.

----